Added validation to check if there was a file uploaded or not
